refactor: Move redux-form utils from entity-detail to tocco-util
- This way, we can use it in other packages as well (for example for
  the search form in the entity-list package)
- formFieldMappings must still be defined in entity-detail (we can't
  move it to tocco-util, because this would result in circular
  dependencies because of the sub grid component which requires the
  entity-list package)
- Only small new feature: `formBuilder` doesn't require an `entity`
   argument anymore (first parameter). If no entity is given, the
   readability check for the entity paths is simply omitted
   (this way we can use it for the search form as well, where we
   won't have any entity data).
