--- packages/entity-detail/src/components/ReduxFormFieldAdapter/index.js
@@ -1,3 +1,0 @@
-import ReduxFormFieldAdapter from './ReduxFormFieldAdapter'
-
-export default ReduxFormFieldAdapter
